refactor: our team section refactoring
