--- src/components/pages/about-us/our-team/team-list/index.ts
@@ -1,3 +1,0 @@
-import TeamList from './team-list';
-
-export default TeamList;
